Changed the condor queue monitoring to wait for jobs only submitted by a particular stage.  This will allow unrelated condor jobs to be in the queue without disturbing the progress of the ogre workflow.
It works by applying a tag to the condor job like:
+ogretag="some_tag"
and running condor_q to search for only jobs with that tag:
condor_q -constraint 'ogretag="some_tag"'

Do to issues passing the nested quote thru rtexec, a wrapper script
is generated in the jobs directory and used to execute condor_q.

This change has been tested for both the convert_ingest and stage_TG (process) 
stages.