Add more smarts to the inserts and updates.  i
For each insert or update, look at the datatypes for each column.
If data type = VARCHAR2, surround the value with single quotes, else
leave alone for numbers.

If the value is a VARCHAR2, but is supposed to be a number, like WINDSPD, then
validate the number, if not a number, set it to zero and issue a warning.  This
is done because NOAO sometimes stuffs error messages in the header keyword
values and we don't want to ingest those.

When checking if an entry exists in LOCATION, use all fields you have, else
just use the ID for checking the existence of entries in the metadata table.

Also I fixed more stuff when parsing headers and filenames.