Use values in header (if they exist) before ingesting into location tables.
The values taken from header are currently:

CCD,BAND,TILENAME

Except for BAND, I'm assuming that the CCD and TILENAME keywords are named
the same as the LOCATION table.  It is FILTER instead of BAND in the header.

Precendence for which keys are inserted into LOCATION table are as follows:

1) Filepath/Filename (Lowest precedence)
2) Fits header
3) Keys provided in filelist (Takes precedence over everything.)