DAA-38 -> DES-48: DAA-8 -> DES-70:
Added generic method to get image ids.  The query is built from the keys
and values of a hashref and results are returned in an array of hashrefs with
each element a hashref with key=id and value = value.

I plan to expand this tremendously to have the returned hash ref array to
contain arbitrary amounts of data, not just IDs.