More bug fixes and design implementation.
Moved where getExposureId is called so that it can make use
of keywords provided in filelist.

Made sure we all of the src files if there are src files in
a filelist.

Call validateHashRef and throw errors of ANY of the
properties that make up a archive structure path
are missing.