Removed check for ingestion (temporarily) to improve performance.
Allow user to enter a batchsize for the ingestions.  The batch object
ingestion will occur after this number of catalogs is read in.  
1000 catalogs is the default.

Filelist can be of any size but will be broken up into batchsize chunks.

Allow user to enter a temporary table name.  If one is not supplied,
TMP_RUN is used.