DES-276:
Removed Big Query from catalog_ingest.pl.  Now the imageid and band are 
retrieved as follows:

my $parentSql = qq{SELECT a.band, b.parentid from image a, catalog b
                   WHERE a.id = b.parentid and b.id = ?};

Made tmptable a required parameter.  Exits if it isn't supplied.

Removed creation of tmp table if it doesn't exist.

Still checking for the existence of the tmp table.  If it is NOT there,
catalog_ingest.pl will issue a level 5 event and exit with a non-zero
status.

Improved error reporting when image, band and catalogid are not found.