DES-289:
Add new columns RA_NUM, DEC_NUM to file ingest as follows:

image table: ra_num is copy of what is in ra, dec_num is copy of what is in dec
exposure table: ra_num is copy of what is in telra, dec_num is copy of what is
in teldec
coadd: ra_num is a copy of what is in telra, dec_num is a copy of what is in
teldec

DES-292:

Make sure crval1 and crval2 values are ingested in RA/DEC values in coadd
table.

DES-293:

When ingesting fpacked src files, don't advance to 2nd hdu.  Needed values are 
in first HDU.