Now works with pgCC and icpc.  I fixed some parts of the code where these compilers issued warnings.  I don't think any of them were actual errors, but it now compiles with -Wall turned on with these compilers.
Updated SConstruct to check for MKL and ACML BLAS and LAPACK libraries.

Also allows ATLAS to use its abbreviated lapack library.

Still need to have it check for CLAPACK, but default of not using 
LAPACK is fine, so no rush there.

I also improved the SmallMatrix and SmallVector classes quite a bit
in response to an error I got with g++ 4.1.  I don't really understand
the error, but the workaround will make these classes a bit faster.
This stuff is still being tested, so test3 might not be successful now, 
but the part of the library that is used for the wl code is ok, so I'm
going to commit now anyway.