DES-383:
More progress on WL catalog ingestions.

Added readWLCatalog to FileUtils.  This reads all WL catalog data and puts it
in a hashRef for batch ingestion.

Added ingestWLHashRef to do the ingestions specific to the WL tables to 
IngestUtils.pm.

Added checkIfWLCatalogIngested so that we don't double-ingest to IngestUtils.pm

Changed wl_catalog_ingest.pl to deal with all of it.