DES-204:
Duplicated some of the functionality of centerpixwcs.pl in file_ingest.pl.  
This is a stopgap solution to get file_ingest.pl to ingest the RA/DEC of the
center pixel in an image.

Calculation is done on red, remap, raw, norm and mask filetypes.  

For coadds:  coadd and coadd_det filetypes.