This tag (v0.40) is the first tag to include the program multishear. It is ready to be tested on the GSN's and then possibly DC5 data.
See:
https://desweb.cosmology.uiuc.edu/confluence/display/WLPipe/WL+Code+and+Data+Products#WLCodeandDataProducts-Multishearprogram
for information on how to run multishear.



Current status:

Works with g++ versions 4.1, 4.3, 4.4.

Works with icpc version 10.1.

Works with icpc version 9 only WITH_OPENMP=false.
  - I am getting segmentation faults with openmp turned on.
  - I suspect this was a bug in icpc 9 that was fixed in version 10.

In all cases, I recommend compiling both tmv and wl with the scons option
WITH_LAPACK=false, especially if you are going to run withopenmp.  
Many lapack distributions are not properly thread-safe, and the native TMV
code is about as fast as lapack, so there isn't much advantage to using
lapack anyway.  So the safest thing to do is compile without it.

The default configuration uses a lot of memory (about 13 GB for a DC4
i-band tile).

To use less memory, use command line option: "multishear_section_size=SIZE"
  - SIZE=10 uses about 700 MB
  - SIZE=20 uses about 2 GB
  - SIZE=30 uses about 4.5 GB
  - In general memory scales roughly as SIZE^2
The trade off, of course, is that using a larger SIZE value is more efficient,
so you should use as large a size as your machine can handle.

NB: The memory usage is dependent on the number of component single-epoch
images.  The numbers above are for i-band.  Less memory would be used for
g,r and more for z.
 


Known problem:

There seems to be a memory leak when running with openmp when using the 
above multishear_section_size option.
  - It's a strage leak, since valgrind doesn't detect it.
  - The symptom is that with each section, top reports more memory usage, 
    but valgrind reports that the memory usage is the same as for the 
    single-threaded run (ie. more or less constant as it progresses through 
    the sections).
  - So it seems to be not a normal memory leak, but rather some kind of 
    problem where the OS is fragmenting the memory or something so that 
    the amount it blocks out for the program keeps growing.
  - The leak didn't occur on my MacBook with either g++ 4.3 or 4.4,
    so this seems to have something to do with the OS.  Maybe the way 
    the OS allocates and releases memory.
  - The leak doesn't seem to happen (or at least was much less severe)
    with icpc 10.1 with a block size of 5.  
    Not sure if this is true for icpc 10.1 with larger block sizes, since
    the computer I tested with 10.1 only has 1 GB of memory.  Also, it could
    be something about this computers OS, rather than icpc 10.1.  Hard
    to say.

Recommended solutions:
  - Use a computer with enough memory to run with one section.
  - Use openmp and hope that your OS doesn't show this symptom. (And then
    watch the memory usage in top to make sure you don't run out.)
  - Use openmp, but set the section size to 5 or lower.
    (And still watch to make sure you don't run out of memory.)
  - Don't use openmp.

I'm still trying to track down what in the code might be causing it, so
hopefully I'll find a way to fix it by the next tagged version.



Current To-do List:

- Implement baackground map method for calculating sky values.  I need to 
  see what kind of format the background map comes in, so this has
  to wait until the DM team has implemented the creation of the maps.
- I suspect it would be faster to calculate the first crude estimate of the 
  size and shape from the coadd image, rather than the se images.  
- Need to calculate the robust singal-to-noise estimate
- I don't think the shear covariance matrix is accurate, so I want to 
  implement a calculation that surveys the local likelihood to calculate
  the covariance.