Fixed some compiling errors when using INST_INT=true.
Skip compilation of functions for <int> that it will never use.
Also specialized some methods for <int> and complex<int> to make sure
they don't call those routines.  If compiled with DEBUG=true, the
compiled code hits a TMVAssert(false).  Perhaps I should change this
to throw an exception rather than use Assert, so it doesn't get turned
off by the (normal) DEBUG=false.  Something like InvalidCallError or 
something like that.  I'd have to add that to the Error heirarchy.

Updated the test suite to work better with <int> tests.  Most of these
had been disabled for int, but with these changes, I can run a larger 
fraction of the test suite on <int> matrices.

Added code to correctly calculate determinants of <int> and complex<int>
matrices.  And added a section in the test suite to test them.

Added a couple items of new functionality:  
m.sumAbs2Elements() and m.addToAll(x).

In SConstruct file, now always use -fno-strict-aliasing for g++.  (It was
just for version <= 4.2.)  I found a bug in g++ 4.4 that was fixed by
turning off strict aliasing.  Since I don't think the strict aliasing
optimizations are probably worth much for TMV, I decided to just always
use the -fno-strict-aliasing flag and be safe.

Also, automatically turn on INST_INT=true if TEST_INT=true.  Similarly, 
automatically turn off TEST_FLOAT or TEST_DOUBLE if the corresponding
INST_FLOAT or INST_DOUBLE = false.  This should automatically avoid
compiling a test suite that won't link with the library.

This is most of the way to the next tagged release, which will be v0.65.