First updated for "script updating" which is the teh abiity to downlokd a new trivial access script, avoiding the lenghty update process which would be needed when jars or jython woudl need updates. The capabluty s to download a new trivialAccess.py, named wiht a version number, and to have a symlink, trivialAccess_current.py point to the most recent downloaded version.
The trivialAccess shell script is chaged to firstl look for trivialAccess_current.py, failing that,
TrivialAceess.py (this supports development...)  the command line is changed to be  like
python file <args> so we do not have to futz with execute permissions and all that when 
we download a script.

Make clean is modified to clean up the symlinks and downloaded files.