OPend to fix robert's isse wiht plots taying alive not havign ti kil them to get the trivial promit. Teh fix is to setup gnuplot, and fix the gnulpot steup to envPrepend.
While at it, had the table file setup trivialaccess. so a setup of 
trivial tools will get you the plotter and trivialAccess.

Make a nont trivial makefile. Put it all in svn.

Took a stab at making a .rst document, (to see what it felt like :-).
so make a doc directory