Fixes requested	for finalcut/coadd production efficiencies.
Code complie, needs test.

runSExtractor.c

added switches to suppliment the command line arguements for the astrOmatic codes.
a number of switched llike --psfex_sex_options have been added.	The string passed
by this	switch is concatentated	to the sex call	in the psfex section of	the codel

Determined by experiment, the astroMAtic codes handle multiple options such that the
rightmost option is the effective option.  This	new capability allows for the passing
of additional options as well as over-riding the options handled in a hard-coded way
in the code itself.

Also Added checks in the command line processing to fail and exit with	non 0
exit code when unknown options are present.

coadd_catalog.c

Added checks of status of sex and psfex commands, if non-0 print status5
and exit with the non-zero status.