Creates a PNG image of all the individual ccds and a mosaic of the whole exposure. It can be run on compressed or uncompressed images. It can also be run on runs that have been finished. If it is ran on finsihed runs, it will copy the compressed image to the /tmp directory, uncompressed it, create the PNG and it will deleted the files from the /tmp directory. If there is a missing ccd, it will fail. This need to be updated.
GeneratePNG.py -l inputlist -d outputdir -b block

The -b block is optional. This can be used in case it is necesary to create png images for a particular
block within the pipeline.
inputlist is a list for one exposure.
For a list of flats, a list for each filter needs to be provided.
The data will be written in outputdir.