describe_table now implements the DES model of having linked data bases. The software now uses oracle meta data to trace synonyms, views and links to find the ALL_TAB_COLS and ALL_COL_COMMENTS  for (I think) any table accessible to the user.
Two new colums ar eprinted, -- the OWNER of the table and the database instance where it is 
defined.

This should make it clear to the cognoscenti when DB links are being traversed to find data. 

Ready for a code review with Todd. 

Want to also add col(s) indicating indexed columns...