auditMD5random.py , checks for a fileclass, that there are correct md5sums in a asample of files, and checks for the correct continuing application of the per-file md5 procedure. (business rules) By default auditMD5random.py   appends to report CSV's /work/QA/techincal/md5sum/
Persistentmetrics.py is a file implementing  a class where I am
experimenting with table design to hold measurement parameters.  Right
now it's using sqllite3 instaniating in memory database -- I'm not
trying to to any harm!.  FWWIW my first try playing with this has me
"loading" relational tables in memory and then using the tables for
report generataion late in the same program.

The table works for "auditing", which is checking that the as-is state
of the system is consistent with business rules.  auditing
is essentails "I measuerd this prpoert of this thign and foung it
to be consistent (or not).  If inconsistnet, explanitory text might
be supplied.  

I need to investigate   measuring /reporting parameters. If of lasting value
this module might need to be in coreutils. If needs to be compared
wiht the schema in the QC faramwork as well.   Just a beginning. 

randomMD5Checker.sh  was removed as an unscalable first attempt at 
auditMD5random.py