Making a version that depends on current operations packages.
Aligning the dist stack to the dev stack will involve some scheduled work
and process, as that is a stack underneath operations, and the whole
project infrastructure has to upgraded with it.  It is a change
that has to be scheduled during some pause in operations, and coordinated
with the Swiss to update desbuild'sinfrastructure, and the software that users download.