Changes to enable unique names for draw QA files.
Now by default the will the take the basename from the output
filename, for that purpose the arg.outname is added to the image
class.

A new command-line option (--draw_basename) allows to pass the
basename for the draw QA png files. 

The default outdir for QA files in how "qa".

This will become tag 0.4.1